requirements.txt
# API and web framework
fastapi>=0.95.0
uvicorn>=0.21.1

# Data processing
pandas>=1.5.3
numpy>=1.24.3

# MinIO client
minio>=7.1.15

# ML libraries
scikit-learn>=1.2.2
joblib>=1.2.0

# MLflow for experiment tracking and model registry
mlflow>=2.3.0

# ONNX for model export
onnx>=1.13.1
skl2onnx>=1.14.0

# Utility libraries
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pydantic>=1.10.7
boto3>=1.26.133